+++ containers/daap-delete-table-for-data-product/src/var/task/delete_table.py
@@ -65,6 +65,10 @@ def handler(event, context):
         return format_error_response(HTTPStatus.BAD_REQUEST, event, str(e))
     except ValueError as e:
         return format_error_response(HTTPStatus.BAD_REQUEST, event, str(e))
+    except Exception as e:
+        message = "Unexpected error while removing schemas. The new version may be in an invalid state"
+        logger.error(message, exc_info=e)
+        return format_error_response(HTTPStatus.INTERNAL_SERVER_ERROR, event, message)
     else:
         msg = f"Successfully removed table '{table_name}'"
         msg += f", data files and generated new matadata version '{new_version}'"