+++ containers/daap-python-base/tests/unit/conftest.py
@@ -8,7 +8,7 @@ from unittest.mock import patch
 
 import boto3
 import pytest
-from moto import mock_glue, mock_s3
+from moto import mock_athena, mock_glue, mock_s3
 
 sys.path.append(join(dirname(__file__), "../", "../", "src", "var", "task"))
 
@@ -68,6 +68,17 @@ def glue_client(region_name):
     """
     with mock_glue():
         client = boto3.client("glue", region_name=region_name)
+
+        yield client
+
+
+@pytest.fixture
+def athena_client(region_name):
+    """
+    Create a mock glue catalogue
+    """
+    with mock_athena():
+        client = boto3.client("athena", region_name=region_name)
 
         yield client
 
@@ -183,19 +194,6 @@ def create_curated_bucket(s3_client, region_name):
         Bucket=os.getenv("CURATED_DATA_BUCKET"),
         CreateBucketConfiguration={"LocationConstraint": region_name},
     )
-
-
-@pytest.fixture
-def create_glue_database(glue_client, data_product_name):
-    glue_client.create_database(DatabaseInput={"Name": data_product_name})
-
-
-@pytest.fixture
-def create_glue_tables(create_glue_database, glue_client, data_product_name):
-    for i in range(3):
-        glue_client.create_table(
-            DatabaseName=data_product_name, TableInput={"Name": f"schema{i}"}
-        )
 
 
 @pytest.fixture