+++ terraform/aws/analytical-platform-management-production/cluster/iam-roles.tf
@@ -3,8 +3,10 @@
 ##################################################
 
 module "iam_assumable_role_cert_manager" {
+  #checkov:skip=CKV_TF_1:Module registry does not support commit hashes for versions
+
   source  = "terraform-aws-modules/iam/aws//modules/iam-assumable-role-with-oidc"
-  version = "5.32.0"
+  version = "5.32.1"
 
   create_role                   = true
   role_name_prefix              = "cert_manager"
@@ -18,8 +20,10 @@ module "iam_assumable_role_cert_manager" {
 ##################################################
 
 module "iam_assumable_role_cluster_autoscaler" {
+  #checkov:skip=CKV_TF_1:Module registry does not support commit hashes for versions
+
   source  = "terraform-aws-modules/iam/aws//modules/iam-assumable-role-with-oidc"
-  version = "5.32.0"
+  version = "5.32.1"
 
   create_role                   = true
   role_name_prefix              = substr("cluster-autoscaler-${module.eks.cluster_id}", 0, 31)
@@ -33,8 +37,10 @@ module "iam_assumable_role_cluster_autoscaler" {
 ##################################################
 
 module "iam_assumable_role_external_dns" {
+  #checkov:skip=CKV_TF_1:Module registry does not support commit hashes for versions
+
   source  = "terraform-aws-modules/iam/aws//modules/iam-assumable-role-with-oidc"
-  version = "5.32.0"
+  version = "5.32.1"
 
   create_role                   = true
   role_name_prefix              = "external_dns"
@@ -48,8 +54,10 @@ module "iam_assumable_role_external_dns" {
 ##################################################
 
 module "iam_assumable_role_external_secrets" {
+  #checkov:skip=CKV_TF_1:Module registry does not support commit hashes for versions
+
   source  = "terraform-aws-modules/iam/aws//modules/iam-assumable-role-with-oidc"
-  version = "5.32.0"
+  version = "5.32.1"
 
   create_role                   = true
   role_name_prefix              = "external_secrets"